requirements.txt
RPi.GPIO
pillow
numpy
spidev
smbus
luma.oled
psutil